set default dtype to float32;add vicuna and fastchat-t5

+++ server.py
@@ -30,7 +30,7 @@ parser.add_argument(
          "`pretrained_path`."
 )
 parser.add_argument(
-    "--dtype", type=str, default="float16",
+    "--dtype", type=str, default="float32",
     help="Dtype to load model."
 )
 parser.add_argument(
